test: Remove the admin visibility test
This is the only test that requires logging in and this functionality
doesn't work out of the box with our new SSO system. It's possible that
we could use an API to log in, but this test is pretty low value so it
doesn't seem worth the effort at this point in time.